Extract MemberPage to separate file
In preparation of a forthcoming commit, the class is being moved to its own
file so it can be required by scraper_test.
